rename discourse_docs to discourse
