+++ canonicalwebteam/discourse/__init__.py
@@ -1,0 +1,9 @@
+from canonicalwebteam.discourse.app import (  # noqa
+    Docs,  # noqa
+    EngagePages,  # noqa
+)
+from canonicalwebteam.discourse.models import DiscourseAPI  # noqa
+from canonicalwebteam.discourse.parsers import (  # noqa
+    DocParser,  # noqa
+    EngageParser,  # noqa
+)
