--- canonicalwebteam/discourse_docs/__init__.py
@@ -1,9 +1,0 @@
-from canonicalwebteam.discourse_docs.app import (  # noqa
-    Discourse,  # noqa
-    EngagePages,  # noqa
-)
-from canonicalwebteam.discourse_docs.models import DiscourseAPI  # noqa
-from canonicalwebteam.discourse_docs.parsers import (  # noqa
-    DocParser,  # noqa
-    EngageParser,  # noqa
-)
