+++ setup.py
@@ -3,7 +3,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="canonicalwebteam.discourse_docs",
+    name="canonicalwebteam.discourse",
     version="2.0.0",
     author="Canonical webteam",
     author_email="webteam@canonical.com",
